backend/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
numpy==1.23.3
pandas==1.5.0
psycopg2==2.8.4